pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "asala"
dynamic = ["version"]
description = "Asala - Cryptographic Content Authenticity Verification"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Asala Contributors"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security :: Cryptography",
    "Topic :: Multimedia",
]
keywords = ["authenticity", "verification", "c2pa", "cryptography", "provenance"]
requires-python = ">=3.8"
dependencies = [
    "cryptography>=3.4.8",
    "cffi<2",
    "c2pa-python>=0.3.0",
    "Pillow>=9.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "scipy>=1.7.0",
    "PyWavelets>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.12.0",
    "flake8>=5.0.0",
    "mypy>=0.991",
    "bandit>=1.7.0",
    "sphinx>=5.0.0",
    "furo",
    "sphinx-copybutton>=0.5.0",
    "sphinxcontrib-mermaid>=0.8.0",
    "pyinstaller>=5.0.0",
    "build>=0.10.0",
]

[project.scripts]
asala = "asala.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/asala"
Documentation = "https://asala.readthedocs.io"
Repository = "https://github.com/your-org/asala"
Issues = "https://github.com/your-org/asala/issues"

[tool.setuptools_scm]
write_to = "python/asala/_version.py"

[tool.setuptools.packages.find]
where = ["python"]

# --------------------------------------------------------------------------
# black
# --------------------------------------------------------------------------
[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# --------------------------------------------------------------------------
# isort — configured to be black-compatible
# --------------------------------------------------------------------------
[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
known_first_party = ["asala"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
src_paths = ["python/asala", "python/tests"]
skip_glob = [".venv/*", "build/*", "dist/*"]

# --------------------------------------------------------------------------
# mypy
# --------------------------------------------------------------------------
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "cv2",
    "cv2.*",
    "pywt",
    "scipy",
    "scipy.*",
    "c2pa",
    "c2pa.*",
]
ignore_missing_imports = true

# --------------------------------------------------------------------------
# bandit
# --------------------------------------------------------------------------
[tool.bandit]
targets = ["python/asala"]
exclude_dirs = [".venv", "build", "dist", "python/tests"]
skips = [
    "B101",  # assert used in tests
    "B311",  # random not used for security
]

# --------------------------------------------------------------------------
# pytest
# --------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["python/tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=asala",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# --------------------------------------------------------------------------
# coverage
# --------------------------------------------------------------------------
[tool.coverage.run]
source = ["python/asala"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
show_missing = true
precision = 2
fail_under = 60

[tool.coverage.html]
directory = "htmlcov"
